requirements.txt
Flask==3.0.3
boto3==1.35.5
mysql-connector-python==9.0.0
